# Optimization of TR and SR lifespan