The purpose of this very simple PR is to improve the management of TR and SR

Tests show that TR and RS bounce a lot in RS network. After being cleared from the cache when their lifespan is reached (at the moment 240 seconds) they come again from somewhere else. We consider them as new and we propagate them again. This can last for hours.

Furthermore sometimes very old TR finally give a positive answer after having bounced everywhere, and a tunnel is opened. Such zombie tunnels don't carry any data and are killed after 1 minute inactivity. In the mean time they will have filled our statistics with garbage and created unneeded traffic between peers.

So in this PR:
- First we increase the lifespan of TR and SR in the turtle cache from 240 to 600 seconds. Keeping known TR & SR in cache during a longer period will help avoid duplicates caused by requests bouncing all over the network. Of course the caches will be initially a little bit bigger, but it does not hurt performance at all.
- Second we drastically reduce the time during which we take into account or forward to our friends the answers we receive to known TR & SR. The previous value of 240 seconds is now reduced to 10 seconds only. Any tunnel result or search result received more than 10 seconds after we put it in cache will be ignored and discarded.