@RetroPooh 2021-02-05
Isn't 10secs too little for forwarding SRs? With 5000 results limit x how many bytes each... how will this fit thru slow bw links like torified nodes?