@csoler 2021-02-05
I'm definitely on RetroPooh side here: 10 secs is really too low. Should be 60. Sometimes TRs gets slowed down for a while and it's too back losing a tunnel opening just because of a timeout. If you increase cache to 600 then 60 will not allow the same bouncing artifacts that we had before. In any way, that needs some testing by different nodes.