@jolavillette 2021-02-06
I don't believe that a TR that needed 60 seconds to come back to us (and still need to reach the node it originates from!!!) will do anything good in a network where RTT is now usually better than 200ms. 
The truth is that these late TR hurt the network a lot.  They are responsible for zombie tunnels. But not only. Each useless TR that we forward instead of dropping it, takes the place of a brand new TR that will be discarded because of the probability system. 
So we save the bad ones and we kill the good ones. We are about to miss a unique chance to do things right for file transfer and distant chat in 0.6.6.